Fix "Inserting a nick by clicking an avatar"
Inserting a nick by clicking an avatar is adding a space to the front of the string.
In the case that it isn't the first word on the line.
If the line is empty, it does the right thing, "%s: ", but if there's text on the line you get " %s". The space should probably be after the nick, if at all.