Fix App freezes on iOS8 when user goes back on Recents from a Room Chat.
Do not call ```destroy``` method during ```dealloc```. Indeed ```destroy``` is used to remove all known references of the current instance (like listeners, delegate...) to ease ```dealloc``` call on the current instance. The allocated resources must be released during ```dealloc``` call, and not on ```destroy``` (So there is no reason to call this method on dealloc).