Read marker - MXKRoomViewController: Improve bubbles table view display.
When the room view controller displays is based on a timeline with an initial event.
The bubble table view content is scrolled by default to display the top of this event at the center of the screen the first time it appears.
We introduce here a new property `centerBubblesTableViewContentOnTheInitialEventBottom` to let the table view center its content on the bottom part of the event. This will be useful to jump to the last unread message and display the marker view.