Read markers should be synchronised across clients
The read marker may be defined on an event which is not displayed in the room history.
We update here the MatrixKit in order to handle this kind of event in the MXKRoomBubbleCellData object.
Previously these events were ignored during the bubbles rendering. Now a bubble component is created even if the event has no actual display.
This component is useful to display the read marker in the room history, and to allow the user to jump on this read marker.

https://github.com/vector-im/riot-meta/issues/8