Fix garbled HTML paragraph syntax during markdown conversion
Fixes vector-im/riot-ios#1500.

It turns out that stripping the leading and ending `<p>`
tags was done in a simplistic way, which ignored the fact
that the string may contain multiple paragraphs.

For example, stripping the leading and ending `<p>` from:

	<p>line 1</p>
	<p>line 2</p>

(incorrectly) used to produce:

	line 1</p>
	<p>line 2

With this patch, we only do leading/ending `<p>` stripping
if the HTML string contains exactly 1 paragraph.

Signed-off-by: Slavi Pantaleev <s.pantaleev@gmail.com>