Do not strip paragraph spacing
This is a continuation of #355 and aims to fix
the rendering of paragraphs.
(related to vector-im/riot-ios#1500)

Stripping paragraph spacing did not work correctly
and caused content like `<p>first</p><p>second</p>`,
which we expected to render as:

```
first

second
```

to actually render as:

```
first
second
```

This completely defeated the utility of paragraphs and turned
longer messages (those having paragraphs) into long
unreadable walls of text.

The code that used to adjust the spacing of paragraph is removed
with this patch, with the hope that detecting and wrapping
bare HTML messages (like `something<br />another`) prevents
them from being turned into (unwanted) paragraphs.

Since all our paragraphs are now wanted, we can let the spacing be
there for all of them.

The only spacing that we want to remove is the bottom-spacing
of the last paragraph, which we still do.

Signed-off-by: Slavi Pantaleev <s.pantaleev@gmail.com>