BF: Push notifications are spontaneously disabling themselves
https://github.com/vector-im/riot-ios/issues/2348

Do not reset MXKAccount.enablePushKitNotifications when trying to refresh pushkit pusher at startup.
If the requests fails, the app will consider that push is disabled whereas there is still a pusher on the homeserver.

This is visible at https://github.com/matrix-org/riot-ios-rageshakes/issues/2618 where the user gets pushes whereas the app setting says there are disabled.

Old logs of https://github.com/matrix-org/riot-ios-rageshakes/issues/2615 show the same thing but on the last app update, the user lost pushes probably because of lack of refresh of the pusher because MXKAccount.enablePushKitNotifications was NO for a long time.