BF: iOS13: Rooms do not load
vector-im/riot-ios/issues/2619

From what I have experienced, [UIView setFrame:] hangs on iOS13 if the view is not part of a view controller and if the viewWillAppear of this view controller is not called.

This means we cannot use anymore a random UITextView within MXKRoomBubbleCellData to compute text height.
I have thus created delegation up to the VC. It computes the height when it can.