MXKRoomDataSource: Fix memory leak in doubly-linked bubble list
Collapsible elements of the `bubbles` array form a doubly-linked list with strong
references stored in `prevCollapsableCellData` and `nextCollapsableCellData`. This
retain cycle prevents the bubbles from being deallocated even when the `bubbles`
array itself is destroyed.

This fix breaks the strong reference cycle by nilifying both `prevCollapsableCellData`
and `nextCollapsableCellData` when the data source is reset.

Note that this commit does *not* fix the very same leak in the `bubblesSnapshot`
array. The latter depends on a race condition though so it is a lot less likely to
occur.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>