Fix collapsing of separately processed events
As reported in vector-im/element-ios#4102, membership changes are sometimes
wrongly collapsed. This happens when events of the same collapsible series
are processed in different batches. For forward directed events, only the
start of a collapsible series is tracked in `collapsingCellDataSeriess`. As
a result, when the event that corresponds to the beginning of the series
was processed in a preceding event batch, the series will be missing in
`collapsingCellDataSeriess` and its summary string won't get updated.

This commit fixes the situation by adding the start of the series to
`collapsingCellDataSeriess` when it is missing and the series is appended
to.

Signed-off-by: Johannes Marbach <n0-0ne+github@mailbox.org>