GH Actions: Fix local MatrixKit.podspec. Just edit it in place
Using the local-podspecs folder did not work because `pod lib lint` cannot find in a relative manner.
We need to use the exact same file name else `pod lib lint` complains.
So, just edit the file in place.

The BuildKite job shows it works there but it is an illusion. The lint command is called with the --quick parameter. It does almost nothing