CI: Introduce GH actions (#804)
* CI: Introduce GH actions

* GH actions: Add MatrixKit.podspec lint

* GH Actions: Fix local MatrixKit.podspec. Just edit it in place

Using the local-podspecs folder did not work because `pod lib lint` cannot find in a relative manner.
We need to use the exact same file name else `pod lib lint` complains.
So, just edit the file in place.

The BuildKite job shows it works there but it is an illusion. The lint command is called with the --quick parameter. It does almost nothing

* GH Actions: Experiment branch selection

* GH Actions: Use fastlane for branch selection

because:
1. Getting the branch from GH actions is hard. It depends if this is a build for a PR or not
2. It is better to abstract the CI tool with fastlane

* GH Actions: Another way for branch selection

* GH Actions: Another way for branch selection

* GH Action: Does global env work?

* GH Actions: Solution found for git_branch